Add EE support for PartitionByPlanNode.
Add EE support for PartionByPlanNode.  The PartitionByPlanNode has an
aggregate operator and a list of partition by expressions.  Its child
node is an order by node which sorts the input table by the partition by
expressions and then by the order by expressions.  The
PartitionByPlanNode does not retain the sort direction, since this is
not relevant to the algorithm.

The output schema is a column for the aggregate operation followed by a
copy of the input schema.  The column index of the aggregate operation
is -1, since it is not copied from the input table.

Note that if pbNode is PartitionByPlanNode, the operation
deserialize(serialize(pbNode)) is not the identity in Java.  This is
because the PartitionByPlanNode has the metadata from the
WindowedExpression.

These are different from previous versions of this plan node.